+++ packages/atomic/src/components/commerce/atomic-commerce-recommendation-interface/atomic-commerce-recommendation-interface.ts
@@ -31,7 +31,7 @@ import {
   createCommerceRecommendationStore,
 } from './store.js';
 
-export type CommerceInitializationOptions = CommerceEngineConfiguration;
+type CommerceInitializationOptions = CommerceEngineConfiguration;
 export type CommerceBindings = CommonBindings<
   CommerceEngine,
   CommerceRecommendationStore,