+++ packages/atomic/src/components/common/smart-snippets/atomic-smart-snippet-feedback-modal/modal-option.ts
@@ -3,7 +3,7 @@ import type {i18n} from 'i18next';
 import {html} from 'lit';
 import type {FunctionalComponent} from '@/src/utils/functional-component-utils';
 
-export interface ModalOptionProps {
+interface ModalOptionProps {
   correspondingAnswer: SmartSnippetFeedback | 'other';
   currentAnswer?: SmartSnippetFeedback | 'other';
   i18n: i18n;