+++ packages/atomic/src/utils/accessibility-utils.ts
@@ -170,9 +170,7 @@ function isFocusable(element: Element) {
   }
 }
 
-export function* getFocusableDescendants(
-  element: Element
-): Generator<HTMLElement> {
+function* getFocusableDescendants(element: Element): Generator<HTMLElement> {
   if (isFocusable(element)) {
     yield element as HTMLElement;
   }